refactor(observability): generalize agent tracing
