+++ README.md
@@ -200,9 +200,8 @@ Placeholder settings sections currently listed in the nav but not fully wired:
 - **Secret redaction**: Tool inputs/log output can be redacted when `SECRET_REDACTION_ENABLED=true`.
 - **Request logging**: Every request gets a request id for log correlation.
 - **Health probes**: `/api/v1/health` is liveness. `/api/v1/health/ready` checks DB connectivity and at least one configured LLM provider.
-- **OpenTelemetry**: Traces include `pawrrtal.turn` and `pawrrtal.llm.chat` spans, model/user/conversation/surface metadata, usage, cost, and Workshop/OpenLLMetry-compatible event hooks.
+- **OpenTelemetry**: Traces include `pawrrtal.turn` and `pawrrtal.llm.chat` spans, model/user/conversation/surface metadata, usage, cost, and OpenLLMetry-compatible event hooks.
 - **Event bus**: A process-local async pub/sub bus publishes turn, scheduled, and agent-response events.
-- **Webhook receiver**: The backend restructure removed the `/webhooks/{provider}` receiver. `WEBHOOK_*` env settings are no-op compatibility placeholders.
 - **Scheduler**: `SCHEDULER_ENABLED=true` starts a cron scheduler. The API can list historical jobs regardless of the flag, but create/delete live jobs require a running scheduler.
 
 ### Desktop shells
@@ -400,9 +399,6 @@ SECRET_REDACTION_ENABLED=true
 STRICT_CONVERSATION_READ_VALIDATION=true
 
 # Ops platform
-WEBHOOK_API_ENABLED=false
-WEBHOOK_API_SECRET=
-GITHUB_WEBHOOK_SECRET=
 SCHEDULER_ENABLED=false
 SCHEDULER_PERSISTENT_JOBSTORE=true
