+++ backend/.env.example
@@ -224,17 +224,7 @@ AUDIT_LOG_RETENTION_DAYS=90
 # STRICT_CONVERSATION_READ_VALIDATION=true
 
 # ============================================================================
-# OPS PLATFORM: WEBHOOKS  (removed compatibility placeholders)
-# ============================================================================
-
-# The backend restructure removed the POST /webhooks routes. These
-# settings remain as no-op placeholders so existing .env files parse.
-# WEBHOOK_API_ENABLED=false
-# WEBHOOK_API_SECRET=
-# GITHUB_WEBHOOK_SECRET=
-
-# ============================================================================
-# OPS PLATFORM: SCHEDULER  (PR #12)
+# OPS PLATFORM: SCHEDULER
 # ============================================================================
 
 # When false the APScheduler lifespan task never starts; the API
@@ -400,15 +390,15 @@ ADMIN_PASSWORD=admin1234
 # /api/v1/workspaces/{workspace_id}/env — see WORKSPACE_ENCRYPTION_KEY above.
 
 # ============================================================================
-# OBSERVABILITY (OpenTelemetry / Raindrop Workshop)
+# OBSERVABILITY (OpenTelemetry)
 # ============================================================================
 # The backend reads OTEL_EXPORTER_OTLP_ENDPOINT from .env automatically.
 # When set, the tracing stack boots and sends OTLP JSON spans.  Unset it
 # and observability is a complete no-op.
 #
-# Local debugging with Raindrop Workshop (https://github.com/raindrop-ai/workshop):
+# Local debugging with an OTLP collector:
 #   curl -fsSL https://raindrop.sh/install | bash
-#   # Workshop listens on :5899.
+#   # The local collector listens on :5899.
 #
 # OTEL_EXPORTER_OTLP_ENDPOINT=http://localhost:5899
 # OTEL_SERVICE_NAME=pawrrtal-backend
