+++ docs/docker.md
@@ -172,9 +172,7 @@ public deploys).
 
 ### Ops platform
 
-Webhooks: `WEBHOOK_API_ENABLED`, `WEBHOOK_API_SECRET`,
-`GITHUB_WEBHOOK_SECRET`. Scheduler: `SCHEDULER_ENABLED`,
-`SCHEDULER_PERSISTENT_JOBSTORE`.
+Scheduler: `SCHEDULER_ENABLED`, `SCHEDULER_PERSISTENT_JOBSTORE`.
 
 ### OAuth
