+++ docs/project-overview.html
@@ -531,7 +531,7 @@
       <p style="margin-top:24px; padding:14px 18px; border:1px solid #d4cfc4; border-radius:8px; background:#f7f4ee; font-size:13px; color:#5a544a;">
         <strong>Snapshot, not source of truth.</strong> This HTML overview is a styled snapshot
         from May 2026 and predates the plugin system (Notion), LCM compaction,
-        the in-process <code>python</code> agent tool, the Workshop / OpenLLMetry
+        the in-process <code>python</code> agent tool, the OpenLLMetry
         OTel integration, and the multi-provider agent loop refactor that replaced
         the Agno SDK. For current architecture see the
         <a href="https://github.com/OctavianTocan/pawrrtal/blob/development/README.md">README</a>
