+++ frontend/content/docs/handbook/index.mdx
@@ -42,7 +42,7 @@ DAG that lets long sessions survive without losing causal links.
   structural changes. Recent ones cover Fumadocs adoption,
   the plugin system + Notion integration, the model-ID format,
   Stagehand vs browser-use, ripping the theming system, and the
-  Workshop OTel integration.
+  agent trace OTel integration.
 - [Agents](/docs/handbook/agents) — how agents should work in this
   repo (`beans` issue tracker, triage flow, domain docs).
 - [Integrations](/docs/handbook/integrations) — how to add a new
